Un-inline full GQ/SQ constructors
